# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DzPulse - Algeria Public Polling Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /polls/{slug}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments on a poll
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Comments list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentListResponse"

    post:
      operationId: createComment
      tags: [comments]
      summary: Create a comment on a poll
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Suggestions
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
...
    Comment:
      type: object
      properties:
        id:
          type: integer
        body:
          type: string
        status:
          type: string
        upvotes:
          type: integer
        createdAt:
          type: string
          format: date-time
        user:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            username:
              type: string
            avatarUrl:
              type: string
              nullable: true
          required: [id, name, username]
        replies:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
      required: [id, body, status, upvotes, createdAt, user, replies]
    CommentListResponse:
      type: object
      properties:
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [comments, total, page, limit, totalPages]
    CreateCommentBody:
      type: object
      properties:
        body:
          type: string
          minLength: 1
          maxLength: 1000
        parentId:
          type: integer
          nullable: true
      required: [body]
...