# lib/api-zod/src/generated/types/createCommentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DzPulse - Algeria Public Polling Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCommentBody {
  /**
   * @minLength 1
   * @maxLength 1000
   */
  body: string;
  parentId?: number | null;
}